Refactor: 이미지 업로드 기능 하위 컴포넌트로 이동
